+++ src/stock_trading/core/views/v1/stock.py
@@ -2,19 +2,20 @@
 # Created by Maximillian M. Estrada on 2024-05-15
 
 from rest_framework import generics
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import IsAdminUser
 
 from core.models import Stock
 from core.serializers import StockSerializer
+from core.permissions import UserReadOnly
 
 
 class StockListView(generics.ListCreateAPIView):
     queryset = Stock.objects.all()
     serializer_class = StockSerializer
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAdminUser | UserReadOnly]
 
 
 class StockDetailView(generics.RetrieveUpdateDestroyAPIView):
     queryset = Stock.objects.all()
     serializer_class = StockSerializer
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAdminUser | UserReadOnly]